Avoiding auto boxing

+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ndarray/BaseNDArrayProxy.java
@@ -39,7 +39,7 @@ public class BaseNDArrayProxy implements java.io.Serializable {
      */
 
     protected int[] arrayShape;
-    protected int length;
+    protected long length;
 	protected char arrayOrdering;
     protected transient  DataBuffer data;
